Server start, stop; Client connect, disconnect implemented
No way to send data yet.

When running via the GameMaker Runner (play button of the GameMaker IDE) make sure to copy the LiteNetLib.dll to the Runner executable directory. Otherwise it won't be found and the program won't work.

This is not a problem when the game is exported because the DLL sits next to the executable, and thus is found.